+++ src/sourcesIndex.js
@@ -1,6 +1,6 @@
 import ReactGA from 'react-ga';
 import doNotTrack from 'donottrack';
-import { setAppName, setVersion, APP_SOURCE_MANAGER } from './config';
+import { setAppName, APP_SOURCE_MANAGER } from './config';
 import { setAppColors } from './styles/colors';
 import routes from './routes/sourceRoutes';
 import initializeApp from './index';
@@ -10,8 +10,6 @@ import initializeApp from './index';
  */
 
 if (!doNotTrack(false)) ReactGA.initialize('UA-60744513-8');
-
-setVersion('3.14.0');
 
 setAppName(APP_SOURCE_MANAGER);